requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.4
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
fonttools==4.51.0
html5lib==1.1
idna==3.7
lxml==5.2.1
oscrypto==1.3.0
pillow==10.3.0
pycparser==2.22
pydyf==0.9.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
xhtml2pdf==0.2.15
zopfli==0.2.3
